requirements.txt
colorama==0.4.4
et-xmlfile==1.1.0
loguru==0.6.0
mysql-connector-python==8.0.28
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.1
protobuf==3.19.4
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
win32-setctime==1.1.0
